requirements.txt
black==24.4.2
blinker==1.8.2
click==8.1.7
flake8==7.0.0
Flask==3.0.3
importlib_metadata==7.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.12.0
Werkzeug==3.0.3
zipp==3.19.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
passlib==1.7.4